Hm, I like Internet a bit better than Web.  Call me old-fashioned?
For context: the original idea was that a javascript on a web page would
be able to add a set of new production rules into your local Avatar Services,
since you can't evaluate the preconditions for a new rule set before you
know what they are.  But the distinction between "by visiting a website
some javascript runs to augment your story model" and "downloading from the
Web" is pretty probably not worth being too anal about.  So "incorporated"
may have been a little more accurate in my original draft, but it's not
worth the awkwardness... at least, not unless I meant to really get into
all the details of how story modules are integrated (and updated, and backed
out)... which the margins of this commit log are too narrow to contain.